chatbot/requirements.txt
python-dotenv==1.0.1
llama-index==0.10.19
llama-index-llms-ollama==0.1.2
fastapi==0.110.0
uvicorn==0.28.0

